Separate credentialsSection file, make cardWrapper more generic
CardWrapper has been made generic to accommodate its use across many
components.
